Refactor project structure and remove unused components
- Remove deprecated animation CSS and related imports
- Delete unused components: AddBookmarkDialog, EditBookmarkDialog, LoadingSpinner, CategoryFilter
- Update layout and page components to remove unnecessary CSS and animation imports
- Simplify component imports and reduce project complexity
- Remove redundant files that are no longer used in the current architecture
