+++ src/app/[mark]/page.tsx
@@ -1,33 +1,35 @@
-import { useTranslations } from "next-intl";
 import "./page.css";
-import { getCategories } from "@/lib/utils";
+import { getBaseUrl, getCategories } from "@/lib/utils";
 import { getBookmarkData } from "@/lib/actions";
-import { DialogCreate } from "@/components/dialog-create";
-
+import { BookmarkUI } from "@/components/bookmark-ui";
 interface BookmarksPageProps {
   params: Promise<{ mark: string }>;
   searchParams: Promise<{ [key: string]: string | undefined }>;
 }
 
-export default async function BookmarksPage({ params, searchParams }: BookmarksPageProps) {
+export default async function BookmarksPage({
+  params,
+  searchParams,
+}: BookmarksPageProps) {
   const { mark } = await params;
   const sp = await searchParams;
+  const status = sp.status;
+  const message = sp.message;
+  let toast: { status: string; message: string } | null = null;
+  if (status && message) {
+    toast = { status, message };
+  }
   const bookmarksData = await getBookmarkData({ mark });
   const categories = getCategories(bookmarksData);
+  const baseUrl = getBaseUrl();
 
   return (
-    // <BookmarksUI
-    //   mark={mark}
-    //   bookmarksData={bookmarksData}
-    //   isLoading={isLoading}
-    //   isDemo={false}
-    //   onDeleteBookmark={handleDeleteBookmark}
-    //   onUpdateBookmark={handleUpdateBookmark}
-    //   onAddBookmark={handleAddBookmark}
-    //   baseUrl={getBaseUrl()}
-    // />
-    <div>
-      <h1>Bookmarks</h1>
-    </div>
+    <BookmarkUI
+      mark={mark}
+      bookmarksData={bookmarksData}
+      categories={categories}
+      toast={toast}
+      baseUrl={baseUrl}
+    />
   );
 }
