--- src/components/category-filter.tsx
@@ -1,56 +1,0 @@
-import { Tag, Check, ChevronDown } from "lucide-react";
-import { useTranslations } from "next-intl";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Button } from "@/components/ui/button";
-
-interface CategoryFilterProps {
-  categories: string[];
-  selectedCategory: string | null;
-  onSelectCategory: (category: string | null) => void;
-}
-
-export function CategoryFilter({
-  categories,
-  selectedCategory,
-  onSelectCategory,
-}: CategoryFilterProps) {
-  const t = useTranslations("BookmarksPage.categories");
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="outline" size="sm" className="gap-1">
-          <Tag className="h-4 w-4" />
-          {selectedCategory ? selectedCategory : t("all")}
-          <ChevronDown className="h-3 w-3 opacity-50" />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="start">
-        <DropdownMenuItem onClick={() => onSelectCategory(null)}>
-          <div className="flex items-center justify-between w-full">
-            <span>{t("all")}</span>
-            {selectedCategory === null && <Check className="h-4 w-4 ml-2" />}
-          </div>
-        </DropdownMenuItem>
-        {categories.map((category) => (
-          <DropdownMenuItem
-            key={category}
-            onClick={() => onSelectCategory(category)}
-          >
-            <div className="flex items-center justify-between w-full">
-              <span>{category}</span>
-              {selectedCategory === category && (
-                <Check className="h-4 w-4 ml-2" />
-              )}
-            </div>
-          </DropdownMenuItem>
-        ))}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
